feat: Add repo module to repos for 5 first methods
